requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
colorama==0.4.4
DAWG-Python==0.7.2
docopt==0.6.2
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
pytest==6.2.3
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
